requirements.txt
Flask==3.0.0
pandas==2.1.3
openpyxl==3.1.2
python-docx==1.1.0
reportlab==4.0.7
openai>=1.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
matplotlib==3.8.2
plotly>=6.3.1
kaleido>=1.1.0


